Changes dpdk tables ternary

+++ src/hardware_dep/dpdk/data_plane/dpdk_tables_ternary.c
@@ -39,7 +39,7 @@
 #ifdef T4P4S_PALMTRIE
     uint8_t* ternary_lookup(lookup_table_t* t, uint8_t* key)
     {
-        (u64)palmtrie_tpt_lookup(t->table, (addr_t)key); //default PALMTRIE_BASIC implemented
+        palmtrie_tpt_lookup(t->table, (addr_t)key); //default PALMTRIE_BASIC implemented
     }
 #else
     uint8_t* ternary_lookup(lookup_table_t* t, uint8_t* key)
